# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Entmoot API
  description: AI-driven site layout automation for real estate due diligence
  version: 0.1.0
paths:
...
  /api/v1/projects:
    get:
      tags:
      - projects
      summary: Get all projects
      description: Retrieve a list of all projects
      operationId: get_all_projects_api_v1_projects_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Get All Projects Api V1 Projects Get
      security:
      - APIKeyHeader: []
    post:
      tags:
      - projects
      summary: Create a new project
      description: Create a new project with configuration and start layout generation
      operationId: create_project_api_v1_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectConfig'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /api/v1/projects/{project_id}/status:
    get:
      tags:
      - projects
      summary: Get project status
      description: Check the status and progress of a layout generation project
      operationId: get_project_status_api_v1_projects__project_id__status_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatusResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/reoptimize:
    post:
      tags:
      - projects
      summary: Re-optimize a project
      description: Re-run optimization with updated configuration parameters
      operationId: reoptimize_project_api_v1_projects__project_id__reoptimize_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - type: object
                additionalProperties: true
              - type: 'null'
              title: Config Updates
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alternatives/{alternative_id}:
    put:
      tags:
      - projects
      summary: Update alternative layout
      description: Update the assets in an alternative layout
      operationId: update_alternative_api_v1_projects__project_id__alternatives__alternative_id__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: alternative_id
        in: path
        required: true
        schema:
          type: string
          title: Alternative Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Update Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Alternative Api V1 Projects  Project Id  Alternatives  Alternative
                  Id  Put
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/results:
    get:
      tags:
      - projects
      summary: Get layout results
      description: Retrieve the generated layout results for a completed project
      operationId: get_layout_results_api_v1_projects__project_id__results_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResults'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Results not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/validate-placement:
    post:
      tags:
      - projects
      summary: Validate single asset placement
      description: Check a single asset position against constraints (for drag-and-drop)
      operationId: validate_placement_api_v1_projects__project_id__validate_placement_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatePlacementRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidatePlacementResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/alternatives/{alternative_id}/export/{export_format}:
    get:
      tags:
      - projects
      summary: Export layout
      description: Export a layout alternative in the specified format
      operationId: export_layout_api_v1_projects__project_id__alternatives__alternative_id__export__export_format__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: alternative_id
        in: path
        required: true
        schema:
          type: string
          title: Alternative Id
      - name: export_format
        in: path
        required: true
        schema:
          type: string
          title: Export Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}:
    delete:
      tags:
      - projects
      summary: Delete a project
      description: Delete a project and its associated data
      operationId: delete_project_api_v1_projects__project_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Project Api V1 Projects  Project Id  Delete
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        error_code:
          type: string
          title: Error Code
          description: Machine-readable error code
          examples:
          - VALIDATION_ERROR
          - PARSE_ERROR
          - STORAGE_ERROR
        message:
          type: string
          title: Message
          description: Human-readable error message
          examples:
          - File validation failed
          - Unable to parse KML file
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
          description: Additional technical details about the error
        timestamp:
          type: string
          title: Timestamp
          description: When the error occurred (UTC)
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Request correlation ID for tracing
        suggestions:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Suggestions
          description: Actionable suggestions for resolving the error
        errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/ErrorDetail'
            type: array
          - type: 'null'
          title: Errors
          description: Detailed field-level errors (for validation)
      type: object
      required:
      - error_code
      - message
      title: ErrorResponse
      description: "Standardized error response model for all API errors.\n\nThis\
        \ model ensures consistent error formatting across the entire API,\nmaking\
        \ it easier for clients to handle errors programmatically.\n\nAttributes:\n\
        \    error_code: Machine-readable error identifier (e.g., 'VALIDATION_ERROR')\n\
        \    message: Human-readable error message\n    details: Optional dictionary\
        \ with additional technical details\n    timestamp: When the error occurred\
        \ (UTC)\n    request_id: Optional request correlation ID for tracing\n   \
        \ suggestions: Optional list of actionable suggestions for resolution\n  \
        \  errors: Optional list of detailed field-level errors"
      example:
        details:
          file_type: kml
          validation_failed: magic_number
        error_code: VALIDATION_ERROR
        errors:
        - code: INVALID_MAGIC_NUMBER
          field: file_content
          message: File signature does not match KML format
        message: File validation failed
        request_id: 550e8400-e29b-41d4-a716-446655440000
        suggestions:
        - Verify the file is a valid KML file
        - Check for file corruption
        timestamp: '2025-11-10T15:30:00Z'
...
    ProjectConfig:
      properties:
        project_name:
          type: string
          maxLength: 255
          minLength: 1
          title: Project Name
          description: Project name
        upload_id:
          type: string
          title: Upload Id
          description: Upload ID from file upload
        assets:
          items:
            $ref: '#/components/schemas/AssetConfig'
          type: array
          minItems: 1
          title: Assets
          description: List of assets to place
        constraints:
          $ref: '#/components/schemas/ConstraintConfig'
          description: Constraint configuration
        road_design:
          $ref: '#/components/schemas/RoadConfig'
          description: Road design configuration
        optimization_weights:
          $ref: '#/components/schemas/OptimizationWeights'
          description: Optimization weights
        dem_upload_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Dem Upload Id
          description: Upload ID for DEM file (optional)
      type: object
      required:
      - project_name
      - upload_id
      - assets
      - constraints
      - road_design
      - optimization_weights
      title: ProjectConfig
      description: Complete project configuration.
    ProjectResponse:
      properties:
        project_id:
          type: string
          title: Project Id
          description: Unique project identifier
        project_name:
          type: string
          title: Project Name
          description: Project name
        status:
          $ref: '#/components/schemas/ProjectStatus'
          description: Current project status
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Project creation timestamp
        message:
          type: string
          title: Message
          description: Response message
      type: object
      required:
      - project_id
      - project_name
      - status
      - created_at
      - message
      title: ProjectResponse
      description: Response model for project creation.
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key